% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{dados}
\alias{dados}
\title{Dados sobre o tratamento de queixas emocionais através de psicoeducação por vídeos}
\format{
Uma tibble com 1196 linhas e as 834 variáveis abaixo:
\describe{
\item{id}{Identificador único}
\item{sexo}{Sex}
\item{idade}{Idade}
\item{aceitou}{Aceite para participar do estudo}
\item{suicidio}{Risco de suicídio}
\item{a_datahora}{Data e hora de avaliação}
}
}
\source{
Projeto de pesquisa TelePSI
}
\usage{
dados
}
\description{
Conjunto de dados que contém informações sobre sujeitos
que passaram pela intervenção de psicoeducação por vídeo
oferecida pelo projeto TelePSI.
}
\keyword{datasets}
