% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar_violin_promis.R
\name{plotar_violin_promis}
\alias{plotar_violin_promis}
\title{Criar gráfico de violino com boxplot de escores da PROMIS no baseline e após 1 mês do tratamento}
\usage{
plotar_violin_promis(dados, dominio)
}
\arguments{
\item{dados}{Tibble ou dataframe com as variáveis a serem plotadas.}

\item{dominio}{Domínio a ser plotado. Pode assumir os valores \code{dep} (depressão), \code{anx} (ansiedade), \code{irr} (irritabilidade) ou \code{sat} (satisfação com a vida).}
}
\value{
Retorna um objeto do ggplot2.
}
\description{
Criar gráfico de violino com boxplot de escores da PROMIS no baseline e após 1 mês do tratamento
}
